testCheckpointFreqBuffered
